LayoutTests/imported/w3c/web-platform-tests/encoding/idlharness-expected.txt

idlharness test

This test validates the WebIDL included in the Encoding Living Standard.


FAIL TextDecoder interface: existence and properties of interface object assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface object length assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface object name assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface: existence and properties of interface prototype object assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface: attribute encoding assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface: attribute fatal assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface: attribute ignoreBOM assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder interface: operation decode(BufferSource,TextDecodeOptions) assert_own_property: self does not have own property "TextDecoder" expected property "TextDecoder" missing
FAIL TextDecoder must be primary interface of new TextDecoder() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL Stringification of new TextDecoder() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL TextDecoder interface: new TextDecoder() must inherit property "encoding" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL TextDecoder interface: new TextDecoder() must inherit property "fatal" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL TextDecoder interface: new TextDecoder() must inherit property "ignoreBOM" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL TextDecoder interface: new TextDecoder() must inherit property "decode" with the proper type (3) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL TextDecoder interface: calling decode(BufferSource,TextDecodeOptions) on new TextDecoder() with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextDecoder"
FAIL TextEncoder interface: existence and properties of interface object assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder interface object length assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder interface object name assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder interface: existence and properties of interface prototype object assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder interface: attribute encoding assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder interface: operation encode(USVString) assert_own_property: self does not have own property "TextEncoder" expected property "TextEncoder" missing
FAIL TextEncoder must be primary interface of new TextEncoder() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextEncoder"
FAIL Stringification of new TextEncoder() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextEncoder"
FAIL TextEncoder interface: new TextEncoder() must inherit property "encoding" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextEncoder"
FAIL TextEncoder interface: new TextEncoder() must inherit property "encode" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextEncoder"
FAIL TextEncoder interface: calling encode(USVString) on new TextEncoder() with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TextEncoder"
